+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Fix [52](https://github.com/babakhani/PersianDate/issues/52)
 - Fix [53](https://github.com/babakhani/PersianDate/issues/53)
 - Fix [56](https://github.com/babakhani/PersianDate/issues/56)
+- Support zero and negative values in input array
 
 ## [1.0.5] - 2018-2-11
 - Fix [35](https://github.com/babakhani/PersianDate/issues/35)